Change the API to query stocks.
Switched from API_DOJO to yfinance package from python. This works and don't need API private/public key for this. The future the key can be deleted, but you can choose not to if want to use the previous API in the future.

I kept the class for API_DOJO, and simply add another for yfinance using pip package.